emitHalideExpr: use isl AST expression generator to print expression
The printed expression is derived from plugging in the inverse schedule.
This inverse schedule may have multiple disjuncts, especially
if the original schedule tree has disjunctive filters, such as
is the case in #200.
The expression can therefore not be assumed to consist of
a single disjunct.

Use the AST expression generator instead of manually trying
to pick the affine expression apart.
Ideally, this should use the AST build at the point
where the user statement was created, but this requires
more invasive changes.

Closes #200