replace findFirstReductionDim by isReductionMember
The original implementation would essentially take any band member
that looks like a reduction and simply take the outermost of those.
Later on, the choice of reduction schedule dimension was made more
stringent to appear at a specific position.
Instead of looking for the first reduction member and then checking
if it is in the right position, simply take the member in the right
position and check if it corresponds to a reduction.