makeSpecializationSet: do not assume parameters have a fixed position
Additionally, do not require the user to add the parameters
to the space first.  Trying to check if a parameter already
appears in a space currently requires functionality that
is only available in the TC copy of isl, which is what this commit
is trying to remove.  Simply adding the parameter if it does not
already appear is much easier.