MapperMemoryPromotionSum2D: drop redundant projection of thread parameters
It is only applied to a set that does not involve any thread parameters.
The projection was introduced in prehistory when the context got included
in the result of "partialScheduleImpl" (without motivation).
This inclusion of the context got removed again later (with motivation),
but the projection in MapperMemoryPromotionSum2D was left in.