AstPrinter::emitStmt: check presence of iterator map for all statements
It is not clear why the check was only being performed for some
statement types.  Perhaps it is not needed for some statement types,
but it is, for example, needed for copy statements and there was
no check for those.
Since an iterator map should have been store for every statement,
simply check it for every statement, even if it is not strictly needed.