emitCudaKernel: avoid relying on isl set variable names in range of schedule
In particular, keep track of the AST build where each AST node was created
and use that to generate AST expressions.
Note that an AST build is a relatively large data structure, so
this may end up consuming some memory if many statements or AST nodes
are involved.
An alternative would be to generate the AST expressions while
the AST is being constructed, as is done in PPCG and in the LLVM
code generator, but this requires two passes over the Halide
data structures, once to generate the AST expressions and
once to print them and would be a more involved change overall.
A third alternative would be to generate AST expressions
for the individual statement indices during the AST generation
(using iteratorMap) and then to generate AST expressions in terms
of those statement indices during the printing of the AST,
plugging in the AST expressions for the statement indices.