split out reduction band after thread mapping to remove nested mapping
mapInnermostBandsToThreads maps coincident band members to threads while
traversing the tree recursively in DFS postorder.  It is possible that
members of two nested bands get mapped to thread dimensions, which is
necessary for parallel reduction support.  In particular, the
reduction detection mechanism splits out the band containing the
reduction loop into a separate band in the tree.  Such mapping is not
valid in general because members of different bands are not
necessarily permutable with each other.  Since thread dimensions are
not executed in any specific order, e.g., there is no guarantee that,
for a given value of threadIdx.y, all threadIdx.x threads will be
executed before (or after) those with another value of threadIdx.y,
permutability is a necessary condition for mapping to multiple thread
dimensions.

Split out the reduction loop immediately before
inserting the reduction synchronization.  During thread mapping, keep
the reduction loop inside its original band.  This avoids the need to
keep track of the relation between the reduction band and its parent,
which can be replaced by the position of the reduction loop in the
band.  As memory promotion may split the band to insert copies, band
splitting must happen before memory promotion passes.  Reduction
synchronization may be inserted at the same time to decrease the
number of passes in the mapping strategy.

Only allow mapping a band to threads if its children are not already
mapped to threads.  The procedure continues recursively so as to
insert all relevant synchronization between children of sequence nodes
even after the thread mapping is complete.