MappedScop::detectReductions: add missing permutability check
If the band with a detect reduction has one or more coincident members,
then it gets split and MappedScop::mapToThreads gets called on both parts.
The call on the outer part would fail to map the coincident members
to thread identifiers because it has a permutability check,
while the handling of reductions assumes that the entire originally
detected reduction band gets mapped to thread identifiers.
Add the same check to MappedScop::detectReductions to prevent this
from happening.
Note that the new check is more strict, because it will also punt
on a single-dimensional non-permutable band mapped to a single
thread identifier.
Also note that this is fairly theoretical because in practice
all bands are permutable.