IterationDomain: store outer loop iterators in an isl::multi_id
In an upcoming commit, affine expressions in terms
of the outer loop iterators will first be constructed
on top of a parameter space and then converted
to an affine expression on top of the iteration domain.
The isl::multi_id is needed in this conversion.
It will also be used inside makeScheduleTreeHelper
to construct the iteration domain.