IterationDomain: also keep track of original parameter space
The function makeIslAffBoundsFromExpr will be changed
to only take a parameter space as input.
Scop::makeIslAffFromStmtExpr therefore needs to have access
to the set of all parameters available at the point where
the iteration domain is constructed.
Store them in IterationDomain for reuse
in Scop::makeIslAffFromStmtExpr.
Note that these parameters are not available as such
in the iteration domain tuple since they are not parameters
from the point of view of the iteration domain.