complete thread mapping in a single band
Since mapping nested bands to threads is invalid, we may complete the
mapping by assigning unused thread dimensions to zero immediately after
mapping coincident and reduction members to threads.  Note that it is
still necessary to insert synchronizations if some band outside the one
mapped to threads has non-coincident members.  The mapping also needs to
be completed before inserting synchronizations between children of a
sequence node, some of which may not be mapped to threads at all.