CudaRTCFunction: do not unload the module that was never loaded
In CudaRTCFunction::Launch, we first add (empty) module to the per-GPU
list of modules, and only then try to load it from PTX. The load part
may fail and throw, the module in the list remaining empty. When the
CudaRTCFunction is destroyed, its destructor tries to unload the module
available in the list. Since the module is empty, unloading fails and
throws. The exception does not seem to be handled, and can even happen
during unwinding of another exception leading to immediate termination.

Load CUmodule and extract CUfunction using local variables, and insert
the values into respective per-GPU lists only when the respective
operation has completed successfully, i.e. without throwing.  If it
throws, nothing is inserted and thus the destructor will not attempt to
clean non-existent objects.