insertCopiesUnder: minor clean-up
The original code uses the arguable incorrectly exported
isl_map_from_domain_and_range with additional wrapping,
which is the same as computing the product of the two inputs.
This latter operation is easier to understand.
The calls are also formatted in a way that emphasizes
the similarities between approximatedRead and writtenElements.