extract out shared Scop::promotedDecl
The promoted declaration information associated to
a given identifier of a promoted tensor reference group
is accessed in two locations, but only one of them
has an explicit check to see if this information is available.
Extract out a promotedDecl that performs the operation (with check)
consistently.
The main motivation, however, is to reduce the clutter inside
mapCopiesToThreads.