promoteToRegistersBelowThreads: use markers to find location below threads
This greatly simplifies the code and is a step towards
the removal of ThreadIdxXScheduleDepthState.

Note that since mapCopiesToThreads inserts such a marker,
while it does not set ThreadIdxXScheduleDepthState,
this means that register promotion will now kick in
on the shared memory copying code, but it will not
actually perform register promotion because it won't find any reuse.
If these checks for reuse turn out to be too expensive,
an explicit check for this situation can be added
to the register promotion later.