promoteToRegistersBelowThreads: extract depth from marker node
The thread mapping filters are now inserted right above
the band members that are mapped to thread identifiers,
while the marker node appears underneath.
The difference in schedule depth is therefore equal
to the number of thread identifiers mapped to a band member.

Once the mapping filter has been changed to explicitly keep
track of the thread identifiers mapped to band members,
this code can be further simplified.

The outermost thread mapping filter is returned because
findThreadMappingAncestor will be reused in isCoalesced,
where the result will also be used to derive the active
domain points outside the mapping.  In this case, it is
important to have a pointer to the outer mapping filter
in order to avoid the mapping filters getting mixed in.