promotionImprovesCoalescing: use markers to find location below threads
This simplifies the code and is the final step towards
the removal of ThreadIdxXScheduleDepthState.
It also avoids the (currently impossible) pitfall of the original code
of the same statement getting mapped differently in different branches.

The code now explicitly handles the case of branches
mapped to a single thread.  It is not entirely clear
how this case was handled before because of the confused
contents of ThreadIdxXScheduleDepthState, but it would
presumably not cause promotion either.