avoid calling "find" on an empty CONDA_PREFIX
In the prehistory, the CONDA_PREFIX variable was introduced in build.sh to
avoid compiling and installing third-party dependences already installed
as conda packages.  However, the check for the libraries being present
is inaccurate as it uses 'find "$CONDA_PREFIX" lib*.so' where
CONDA_PREFIX may be (and in non-conda builds is) empty.  In the case of
empty path, "find" searches for the given files in recursively starting
from the current directory.  When called after a previous successful
build, "find" finds the libraries in their _build_ directories and skips
the following compilation passes, including checks for changes,
entirely.  It also does not install the libraries if they were removed.

Check that CONDA_PREFIX is not empty before using find to detect
pre-installed libraries.