Replace C2 TcOp::profile_ by TcOp::checkSizes_
C2 has its own way of profiling, this flag is unnecessary.
Instead C2 needs to be able to check sizes so add that..

Note that on a size mismatch it is likely a CHECK will  terminate the progra.
I opened #351 to address that part.

Still once #307 has landed there will be a need for memoization of
OptionsCacheKey -> TcExecutor if you want a TcOp to work across multiple sizes
and until we have some end-to-end parametric code generation.