[C++ API] Step 1: drop manual and compilation caches
This is the first PR towards implementing the API proposed in #307.
As part of refactoring the OptionsCache we realized that compilation caches may
have been quite useful in the beginning but are a premature optimization in the
new API.

Removing them has multiple benefits such as:
1. simplify the codebase CRTP + inheritance + custom traversals and
serialization are not fun to maintain
2. simplify porting to the new API where ExecutionEngine is dead
3. caches seem to add overhead in the context of cooperative groups
4. in the new API we can cache executors which removes the need to rely on
compilation caches
5. tuning and options caches is the expensive bit that needs to be memoized

This seems corroborated by the [discussion](https://github.com/facebookresearch/TensorComprehensions/pull/339#discussion_r183937583).

Note that some tests in test_tc_mapper need to be temporarily degraded because
we do not have a simple way to extract generated code before the new API.