[C++ API] Step 4: add the new ATen compiler files
This PR adds the files implementing the ATen user-facing part (#324) on
top of the new compiler API (#322). The ExecutionEngine has been retired
and the user-facing API reflects this.

This provides a minimal API to wrap tc/core/compiler.h under a tensor
library like ATen. In particular this gets rid of the ATenCompilationUnit
class which kept state about parsed TCs.

Tinkering with the underlying TreeRef memoization is penny-wise and pound
foolish: parsing overhead never matters and it adds cognitive overhead to the
API. Therefore this API only provides functionality operating on TC strings.

To guard against overhead, the executors themselves should be memoized and
reused.

As part of the bigger refactoring, these files are not yet activated
(or even compiled) therefore the changeset cannot be tested independently.
This new implementation will be switched in as part of the last commit of the
global refactoring.