[CircleCI] use matching names for storing and restoring ATen cache
7b36d6d210fa16231f5df42744a821320b2cb171 (Sun Apr 22 08:54:09 2018
-0700) replaced ATen submodule with PyTorch, updating the test script
instructions.  On build-1404 instances, it changed the compilation cache
key prefix from v1-aten to v1-pytorch when restoring, but did not change
the prefix for storing the compiled files.  As a result, ATen is being
recompiled on every test run unnecessarily.  Use v1-aten as key prefix
in both storing and restoring the compilation cache.  -aten is chosen
over -pytorch for consistency with build-1604.